Add support for downloading Instagram videos and photos (Issue #216)
This commit adds support for downloading videos and photos from Instagram. The new feature uses the yt-dlp program to download the videos and photos in MP4 format. The videos and photos are saved to the current directory or to the directory specified by the dir variable.

The commit also adds a new method called downloadFromInstagram() to the FileDownloader class. This method takes a link to an Instagram video or photo and a directory as arguments and downloads the video or photo to the specified directory.